Rename themeConfig stub file
