+++ src/stubs/build.js
@@ -1,4 +1,4 @@
 import { build } from '@artrz/thunderbird-theme-builder';
-import theme from './themeColors.js';
+import theme from './themeConfig.js';
 
 build(theme, { stylesheet: 'themeCustomStyles.scss' });
